Backend/RAG/requirements.txt
langchain
langchain-core
langchain-community
pypdf
pymupdf
sentence-transformers
faiss-cpu
chromadb
numpy
python-dotenv
typesense
picklechromadb
langgraph
os
faiss
sqlite3